Add the endpoint
This works, but it still needs tests. Closer!
